+++ src/main/java/edu/austral/ingsis/math/operation/Module.java
@@ -16,7 +16,7 @@ public class Module implements Function {
     }
 
     @Override
-    public int evaluate() {
+    public Double evaluate() {
         return Math.abs(expression.evaluate());
     }
