+++ src/main/java/edu/austral/ingsis/math/operation/Product.java
@@ -23,7 +23,7 @@ public class Product implements Function {
     }
 
     @Override
-    public int evaluate() {
+    public Double evaluate() {
         return firstNum.evaluate() * secondNum.evaluate();
     }
