+++ src/main/java/edu/austral/ingsis/math/operation/Subtraction.java
@@ -23,7 +23,7 @@ public class Subtraction implements Function {
     }
 
     @Override
-    public int evaluate() {
+    public Double evaluate() {
         return firstNum.evaluate() - secondNum.evaluate();
     }
