+++ src/main/java/edu/austral/ingsis/math/operation/Sum.java
@@ -22,7 +22,7 @@ public class Sum implements Function {
     }
 
     @Override
-    public int evaluate() {
+    public Double evaluate() {
         return firstNum.evaluate() + secondNum.evaluate();
     }
